+++ dynts/dsl/functions/__init__.py
@@ -2,4 +2,4 @@ import simple
 import simplexy
 
 
-FUNCTIONS = ['log', 'delta', 'ldelta', 'ma', 'max', 'med', 'min', 'zscore', 'prange', 'avol']
+#FUNCTIONS = ['log', 'delta', 'ldelta', 'ma', 'max', 'med', 'min', 'zscore', 'prange', 'avol']