+++ dynts/dsl/functions/simple.py
@@ -15,68 +15,154 @@ class ScalarFunction(FunctionBase):
         
 
 class Log(ScalarFunction):
-    """Delta"""
+    """\
+
+log
+========
+
+Calculate the natural logarithm of a timeseries. It applies to each
+value and return a timeseries with exactly the same dimensions.
+
+.. math::
+
+    \log{(ts)}
+"""
     def apply(self, ts, **kwargs):
         return ts.log(**kwargs)
     
     
 class Delta(ScalarFunction):
-    """Delta"""
+    """\
+
+delta
+===========
+
+First order difference evaluated as
+
+.. math::
+
+    \delta y_t = y_t - y_{t-1}
+
+"""
     def apply(self, ts, **kwargs):
         return ts.delta(**kwargs)
     
     
 class LDelta(ScalarFunction):
-    """Log-Delta. Can be used for calculating percentage moments."""
+    """\
+
+ldelta
+============
+
+Calculate the log-delta of a timeseries. This is the first order difference
+in log-space useful for evaluationg percentage moments
+    
+.. math::
+
+    {ldelta}(y_t) = \log \frac{y_t}{y_{t-1}}
+
+"""
     def apply(self, ts, **kwargs):
         return ts.logdelta(**kwargs)
 
 
 class Ma(ScalarFunction):
-    """Moving average function"""
+    """\
+
+ma
+=============
+
+Arithmetic moving average function.
+"""
     def apply(self, ts, **kwargs):
         return ts.rollmean(**kwargs)
     
     
 class Max(ScalarFunction):
-    """Moving max function"""
+    """\
+
+max
+=============
+
+Moving max function.
+"""
     def apply(self, ts, **kwargs):
         return ts.rollmax(**kwargs)
 
 
 class Med(ScalarFunction):
-    """Moving median function"""
+    """\
+    
+med
+=============
+
+Moving median function.
+"""
     def apply(self, ts, **kwargs):
         return ts.rollmedian(**kwargs)    
     
     
 class Min(ScalarFunction):
-    """Moving min function"""
+    """\
+
+min
+===========
+
+Moving min function.
+"""
     def apply(self, ts, **kwargs):
         return ts.rollmin(**kwargs)
     
     
 class zscore(ScalarFunction):
-    """Rolling Z-Score function"""
+    """\
+    
+zscore
+=============
+
+Rolling Z-Score function:
+
+.. math::
+
+    zs = \frac{x_n - x_{n-w}}{\sigma_n}
+    
+"""
     def apply(self, ts, **kwargs):
         return tsfunctions.zscore(ts, **kwargs)
     
 
 class prange(ScalarFunction):
-    """Rolling Percentage range function"""
+    """\
+    
+prange
+=========
+
+Rolling Percentage range function.
+"""
     def apply(self, ts, **kwargs):
         return tsfunctions.prange(ts, **kwargs)
     
     
     
 class Avol(ScalarFunction):
-    '''Annualised volatility'''
+    '''\
+    
+avol
+=======
+
+Annualised volatility.
+'''
     def apply(self, ts, **kwargs):
         return ts.rollavol(**kwargs)
     
     
-class Regression(FunctionBase):
-    """Calculate the **linear regression** of one series with respect
+class reg(FunctionBase):
+    """\
+
+regr
+==========
+
+Calculate the **linear regression** of one series with respect
 to one or more series. For example::
 
     regr(GOOG,YHOO)
@@ -85,11 +171,12 @@ will calculate
 
 .. math::
 
-    y_i = \beta x_i + \alpha
+    y_i = b x_i + a
     
 There are two optional parameters:
 
-* *alpha* default is ``1``. If set to zero alpha won't be included in the regression."""        
+* *alpha* default is ``1``. If set to zero alpha won't be included in the regression.
+"""        
     name = 'regr'
     
     def __call__(self, input, **kwargs):